Add playwright tests
